INDY-1303: add fail test with advancing node
Changes:
- added node with advancing node. View change start in ordering and one node break due to advance.
